requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.7
django-extensions==1.6.7
django-toolbelt==0.0.1
djangorestframework==3.3.3
flatlib==0.2.1
gnureadline==6.3.3
gunicorn==19.6.0
ipdb==0.10.1
ipython==4.2.0
ipython-genutils==0.1.0
pexpect==4.1.0
pickleshare==0.7.2
psycopg2==2.6.1
ptyprocess==0.5.1
pyswisseph==2.0.0.post2
requests==2.12.5
simplegeneric==0.8.1
six==1.10.0
static3==0.7.0
traitlets==4.2.1
whitenoise==3.2
